Move payout_curve under payouts module
