+++ model/src/lib.rs
@@ -30,7 +30,7 @@ mod contract_setup;
 pub mod hex_transaction;
 pub mod libp2p;
 pub mod olivia;
-pub mod payout_curve;
+pub mod payouts;
 mod rollover;
 
 pub use cfd::*;
