+++ model/src/payouts.rs
@@ -1,0 +1,3 @@
+mod payout_curve;
+
+pub use payout_curve::calculate;
