+++ model/src/payouts/payout_curve.rs
