+++ model/src/payouts/payout_curve/README.md
