+++ model/src/payouts/payout_curve/compat.rs
