+++ model/src/payouts/payout_curve/utils.rs
@@ -1,4 +1,4 @@
-use crate::payout_curve::Error;
+use crate::payouts::payout_curve::Error;
 use ndarray::prelude::*;
 use ndarray::s;
 use std::cmp::Ordering;
